Added numerous missing RFCs to the list
Summary:
Further to T48.
In this batch are mostly RFCs which are explicity mentioned in the code
or listed in the old specs.html document.  An effort has been made to
ensure that the RFCs added are actually relevent to the project.  Some
re-ordering has occurred, in an effort to keep the list sorted, which
appeared to be the desired result.

There were some RFCs already in the list which have been obsoleted or
updated by subsequent issues.  These have been noted following existing
convention.  In some cases I've added in obsolete RFCs where doing so
presents expected RFC numbers as supported, but where seminal RFCs had
not previously been listed.

Test Plan: None

Reviewers: #documentation_reviewers, vanmeeuwen

Reviewed By: #documentation_reviewers, vanmeeuwen

Maniphest Tasks: T48

Differential Revision: https://git.cyrus.foundation/D30